Metal parity: KV caches as per-size states, chunk-128 prefill, lean decode path
Export (gemma_chat):
- New MIL pass global_kv_caches_to_states: post-materialization, the 6
  global KV caches become fp16 state features per concrete-size function
  (read_state at entry, coreml_update_state at exit; the fill_like+add
  write is load-bearing — a bare slice_update write silently loses state
  on macOS 26). sliding_pos_ring stays int32 I/O.
- Fix materialize weight-id collisions for same-named consts across
  functions (invisible at chunk 8; corrupted the package at chunk 128).
- CHUNK_SIZE 8 -> 128 (+ guard: materialize sizes must be >= chunk).
- Logit projection is no longer quantized: int8 constexpr weights feeding
  a matmul are constant-folded by MPSGraph on every first prediction per
  process (~18 s, 27 GB transient, uncacheable; int4 is never folded).
  The traced logit matmul now runs in fp16 with a plain fp16 const
  (+402 MB) and casts to fp32 after; argmax-identical on the isolated
  change, and strictly more accurate than int8.

Runtime (GemmaCore):
- Per-size MLState (sliding + global caches) created from each size
  pair's model handle; growth migrates state old->new with one
  front-aligned prefix copy per cache. Global caches never cross the
  prediction boundary again (the per-step I/O copies were the decode
  bottleneck at large contexts).
- MPSGraph specialization is absorbed at ensureLoaded/growth time via a
  throwaway prediction into a scratch state, off the measured path.
- fp16 logits end to end with preallocated IOSurface/page-aligned
  outputBackings, double-buffered; reused input scalars.
- Decode growth headroom 1024 -> 256.
- Sampling: vDSP + bounded top-k preselect with exact full-sort
  fallback; 22.5 ms -> 0.28 ms per token, bit-identical in 536/540
  trials (rest are exact ties).
- Old-format models are rejected with a clear re-export message.

Measured (M4 Pro, cpu-and-gpu, 3 runs/cell vs pre-change baseline):
- prefill: 86 -> 740 tok/s at ctx 512 (8.6x); 62 -> 639 at 8192 (10.3x)
- decode at 8192: 2.1 -> 8.7 tok/s (4.2x); flat-ish across context now
- decode W/token at 8192: 5.13 -> 0.86
- decode at short ctx: 12.5-12.8 vs 15.8-17.3 baseline — the baseline
  number bundled 22 ms/token of old sampling into fewer, slower steps;
  per-step model time regressed ~36 -> ~80 ms and is the next target.
- CLI multi-turn chat verified: memory, /reset, coherence; 12.2 tok/s.

+++ gemma_chat/decode_coreml.py
@@ -350,8 +350,13 @@ def decode_step(
         x = x * lp['layer_scalar']
 
     x = _rmsnorm(x, params['norm']['scale'])
-    logits = jnp.dot(x[0, 0].astype(jnp.float32),
-                     params['embed_tokens'].T.astype(jnp.float32))  # (vocab,)
+    # fp16 matmul, fp32 only from the logits onward.  Upcasting the weight
+    # instead would put a [dim, vocab] *fp32* constant in the graph — 1.6 GB
+    # that no pass can shrink, since the export leaves this tensor unquantized
+    # (see ``mil_passes/quantize_const_weights``: int4 cannot carry the logits
+    # and int8 is what MPSGraph constant-folds on every first prediction).
+    logits = jnp.dot(x[0, 0],
+                     params['embed_tokens'].T).astype(jnp.float32)  # (vocab,)
     if cfg.final_logit_softcap is not None:
         cap = cfg.final_logit_softcap
         logits = jnp.tanh(logits / cap) * cap
@@ -614,8 +619,10 @@ def chunk_prefill_step(
         x = x * lp['layer_scalar']
 
     x = _rmsnorm(x, params['norm']['scale'])
-    logits = jnp.dot(x[0].astype(jnp.float32),
-                     params['embed_tokens'].T.astype(jnp.float32))  # (C, vocab)
+    # fp16 matmul, fp32 from the logits onward — see the note in
+    # :func:`decode_step` on why the weight must not be upcast.
+    logits = jnp.dot(x[0],
+                     params['embed_tokens'].T).astype(jnp.float32)  # (C, vocab)
     if cfg.final_logit_softcap is not None:
         cap = cfg.final_logit_softcap
         logits = jnp.tanh(logits / cap) * cap
